requirements.txt
Faker==0.7.17
Flask==0.12.2
Flask-JWT==0.3.2
Flask-Login==0.4.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.2
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
PyJWT==1.5.2
SQLAlchemy==1.1.11
SQLAlchemy-Utils==0.32.14
Werkzeug==0.12.2
alembic==0.9.2
aniso8601==1.2.1
apipkg==1.4
certifi==2017.4.17
chardet==3.0.4
click==6.7
execnet==1.4.1
factory-boy==2.8.1
flask-marshmallow==0.8.0
freezegun==0.3.9
idna==2.5
itsdangerous==0.24
marshmallow==2.13.5
marshmallow-sqlalchemy==0.13.1
pep8==1.7.0
psycopg2==2.7.1
py==1.4.34
pytest==3.1.2
pytest-cache==1.0
pytest-pep8==1.0.6
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2017.2
requests==2.18.1
six==1.10.0
urllib3==1.21.1
